# Conflicts:
#	CHANGELOG.md
#	stubs/inertia-react/resources/js/Components/Input.js
#	stubs/inertia-react/resources/js/Pages/Auth/Login.jsx
#	stubs/inertia-react/resources/js/Pages/Auth/Register.js

+++ src/Console/InstallsApiStack.php
@@ -9,7 +9,7 @@ trait InstallsApiStack
     /**
      * Install the API Breeze stack.
      *
-     * @return void
+     * @return int|null
      */
     protected function installApiStack()
     {
@@ -17,10 +17,10 @@ trait InstallsApiStack
 
         // Controllers...
         $files->ensureDirectoryExists(app_path('Http/Controllers/Auth'));
-        $files->copyDirectory(__DIR__.'/../../stubs/api/App/Http/Controllers/Auth', app_path('Http/Controllers/Auth'));
+        $files->copyDirectory(__DIR__.'/../../stubs/api/app/Http/Controllers/Auth', app_path('Http/Controllers/Auth'));
 
         // Middleware...
-        $files->copyDirectory(__DIR__.'/../../stubs/api/App/Http/Middleware', app_path('Http/Middleware'));
+        $files->copyDirectory(__DIR__.'/../../stubs/api/app/Http/Middleware', app_path('Http/Middleware'));
 
         $this->replaceInFile('// \Laravel\Sanctum\Http', '\Laravel\Sanctum\Http', app_path('Http/Kernel.php'));
 
@@ -32,10 +32,11 @@ trait InstallsApiStack
 
         // Requests...
         $files->ensureDirectoryExists(app_path('Http/Requests/Auth'));
-        $files->copyDirectory(__DIR__.'/../../stubs/api/App/Http/Requests/Auth', app_path('Http/Requests/Auth'));
+        $files->copyDirectory(__DIR__.'/../../stubs/api/app/Http/Requests/Auth', app_path('Http/Requests/Auth'));
 
         // Providers...
-        $files->copyDirectory(__DIR__.'/../../stubs/api/App/Providers', app_path('Providers'));
+        $files->copyDirectory(__DIR__.'/../../stubs/api/app/Providers', app_path('Providers'));
+        $this->replaceInFile("HOME = '/home'", "HOME = '/dashboard'", app_path('Providers/RouteServiceProvider.php'));
 
         // Routes...
         copy(__DIR__.'/../../stubs/api/routes/api.php', base_path('routes/api.php'));
@@ -56,21 +57,22 @@ trait InstallsApiStack
             copy(base_path('.env.example'), base_path('.env'));
         }
 
-        $this->replaceInFile(
-            'APP_URL=http://localhost',
-            'APP_URL=http://localhost'.PHP_EOL.'FRONTEND_URL=http://localhost:3000',
-            base_path('.env')
+        file_put_contents(
+            base_path('.env'),
+            preg_replace('/APP_URL=(.*)/', 'APP_URL=http://localhost:8000'.PHP_EOL.'FRONTEND_URL=http://localhost:3000', file_get_contents(base_path('.env')))
         );
 
         // Tests...
-        $this->installTests();
+        if (! $this->installTests()) {
+            return 1;
+        }
 
         $files->delete(base_path('tests/Feature/Auth/PasswordConfirmationTest.php'));
 
         // Cleaning...
         $this->removeScaffoldingUnnecessaryForApis();
 
-        $this->info('Breeze scaffolding installed successfully.');
+        $this->components->info('Breeze scaffolding installed successfully.');
     }
 
     /**
@@ -84,7 +86,7 @@ trait InstallsApiStack
 
         // Remove frontend related files...
         $files->delete(base_path('package.json'));
-        $files->delete(base_path('webpack.mix.js'));
+        $files->delete(base_path('vite.config.js'));
 
         // Remove Laravel "welcome" view...
         $files->delete(resource_path('views/welcome.blade.php'));
